Fix hanging tests on OSX
The test hangs were caused by the flock() attempt in Pod::Tidy.
The flock() fails on OSX because temp files created by File::Temp
lock their files if locking is available. On linux this doesn't
seem to matter because LOCK_EX is not available. On darwin
LOCK_EX is defined and the temp file is locked. Pod::Tidy opens
the file and attempts to get a lock but this fails because
the temp file still has the lock. This fix here is to explicitly
release the lock before passing the temp file name to Pod::Tidy.